Iter4 (#4)
* implement server

* reduce mod ver

* transform constants to camelcase

* fix erorr cases

* implement collector & client

* add test for update handler

* rm unused types

* clean

* clean&undo

* retur missed server.go

* add chi; add root handler

* fix float

* WIP

* remove acc from storage; use sum in collector

* WIP

* undo precision; set it to 3

* use sum in storage

* WIP

* WIP

* fix missed return

* rm wg

* format float with min possible precision

* rm unused dep

* add flags for server & agent

* edit agent regex

* rm unused mux

* edit addr validation

+++ internal/collector/collector.go
@@ -12,7 +12,7 @@ type Mtcs struct {
 
 type Collector struct {
 	mtcs         Mtcs
-	pollInterval uint8
+	pollInterval uint
 }
 
 func New(pollInterval uint8) *Collector {
